requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.8
#cloudinary==1.42.2
colorama==0.4.6
databases==0.9.0
dnspython==2.7.0
#email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycountry==24.6.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyodbc==5.1.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
request==2.32.4
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.41.3
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.30.3
watchfiles==1.0.3
websockets==14.1